Removed the details and summary sections
There is a bug in how ie 11 renders the expandable detail summarys.

It doesn't.

So instead of tracking down what was the issue because it wasn't immediately obvious, I've removed them converted the page to a single page and then added a table of contents to the beginning as per the Canada.Ca style guide design pattern for an in page table of contents

https://design.canada.ca/common-design-patterns/in-page-toc.html